Addressing feedback for the migrations (#104104)
